Rename SyncUser::identity() to SyncUser::user_id()
The word "identity" was used for two different concepts in SyncUser:
`identity()`` returned the server-assigned unique id for the User, while
`identities()` returned information on the credentials used to log into the
user. This was very confusing, so rename the first one to user_id (and update
all the other places where identity was used for that).